refactor: 영화 관련 로직과 상태를 MoviesContainer 컴포넌트로 분리
- 영화 썸네일과 영화 목록 렌더링을 MoviesContainer 컴포넌트에서 처리하도록 관심사 분리
- App 컴포넌트 내 movies, isLoading, page, searchPage, show 상태를 MoviesContainer로 이동하여 App 컴포넌트의 복잡성 완화
