+++ src/main.ts
@@ -2,5 +2,5 @@ import App from "./App.js";
 
 const app = new App();
 (async function () {
-  await app.init();
+  await app.render();
 })();
